requirements.txt
# Django and Django-related packages
Django>=5.1.5
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-filter>=23.5
django-modeltranslation>=0.18.11
django-cors-headers>=4.3.1
django-environ>=0.11.2
django-sendgrid-v5>=1.2.2
django-auditlog>=2.3.0

# Database
psycopg2-binary>=2.9.9

# PDF Generation
reportlab>=4.0.8

# Payment Gateway
razorpay>=1.4.1

# Utilities
python-dateutil>=2.8.2